refactor: 구 색상 토큰 사용처를 tokensV2로 전환
- 구 tokens/color.css를 쓰던 10개 파일을 tokensV2/color.css로 전환. 전부 동일 hex 값의 primitive로 매핑해 시각 변화 없음 (a 계열은 gray999_50→gray999_a50 형식으로 이름 변경, primary→purple600, primary_dark→purple700)
- gray999_20(딤 0.2)은 공식 디자인 토큰에 없는 값이라 토큰을 신설하지 않고, 유일한 사용처인 CommunityComingSoonModal backdrop을 v2 Popup과 같은 리터럴 방식으로 전환
- 사용처가 사라진 구 tokens/color.css.ts 삭제
